feat(native): R3 field hardening — offline queue, OTA updates, lead editing
Offline (audit #2, #10): onlineManager wired to NetInfo — mutations fired
offline now PAUSE instead of failing; Mark Delivered queues, survives app
restarts (mutation defaults + persister + resumePausedMutations on
hydrate), and replays on reconnect. Global amber offline banner. Driver
sees 'Saved offline — will sync when back on network'.

OTA updates (audit #6): expo-updates installed + EAS Update configured
(channels per profile, runtimeVersion appVersion). On launch the app
checks, fetches and reloads JS updates — old sideloaded APKs finally get
updates without reinstalling; only native-module changes need a new APK.

Lead editing (audit #9): QuickActionsModal extracted to
src/components/LeadQuickActions and mounted on the lead DETAIL screen
(⚡ Edit next to Call/WhatsApp) — no more backing out to the list to
change status/stage/follow-up.

Verified: native tsc clean + expo export bundles.
