+++ apps/native/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@maiyuri/shared": "file:../../packages/shared",
     "@react-native-async-storage/async-storage": "2.2.0",
+    "@react-native-community/netinfo": "11.4.1",
     "@sentry/react-native": "~7.2.0",
     "@supabase/supabase-js": "^2.47.12",
     "@tanstack/query-async-storage-persister": "^5.101.2",
@@ -29,6 +30,7 @@
     "expo-router": "~6.0.0",
     "expo-secure-store": "~15.0.0",
     "expo-status-bar": "~3.0.0",
+    "expo-updates": "~29.0.18",
     "nativewind": "^4.2.1",
     "react": "19.1.0",
     "react-dom": "19.1.0",
