+++ apps/native/src/hooks/use-deliveries.ts
@@ -31,10 +31,16 @@ export type CompleteDeliveryInput = {
   notes?: string;
 };
 
-/** POST /api/deliveries/:id/complete — proof-of-delivery. */
+/**
+ * POST /api/deliveries/:id/complete — proof-of-delivery.
+ * mutationKey matches the default in query-client.ts so an offline
+ * "Mark Delivered" queues (pauses), survives app restarts via the
+ * persister, and replays automatically when the network returns.
+ */
 export function useCompleteDelivery() {
   const queryClient = useQueryClient();
   return useMutation({
+    mutationKey: ['complete-delivery'],
     mutationFn: ({ id, ...body }: CompleteDeliveryInput) =>
       api.post(`/api/deliveries/${id}/complete`, body),
     onSuccess: () => {
